[fix] test_add_group_members

+++ tests/test_util.c
@@ -84,14 +84,14 @@ bool is_equal_data(ProtobufCBinaryData *data1, ProtobufCBinaryData *data2) {
 }
 
 bool is_equal_str(char *str1, char *str2) {
-    int len1 = strlen(str1);
-    int len2 = strlen(str2);
+    size_t len1 = strlen(str1);
+    size_t len2 = strlen(str2);
 
     if (len1 != len2) {
         return false;
     }
 
-    int i;
+    size_t i;
     for (i = 0; i < len1; i++) {
         if (str1[i] != str2[i]) {
             return false;
